refactor: change component casing
